+++ pymove/preprocessing/compression.py
@@ -90,8 +90,8 @@ def compress_segment_stop_to_point(
             )
 
         print('...setting mean to lat and lon...')
-        lat_mean = np.full(move_data.shape[0], -1.0, dtype=np.float32)
-        lon_mean = np.full(move_data.shape[0], -1.0, dtype=np.float32)
+        lat_mean = np.full(move_data.shape[0], -1.0, dtype=np.float64)
+        lon_mean = np.full(move_data.shape[0], -1.0, dtype=np.float64)
 
         if drop_moves is False:
             lat_mean[move_data[~move_data[label_stop]].index] = np.NaN
